+++ dsi/offline/acceptance/README.md
@@ -1,0 +1,1 @@
+This directory includes the code used for estimating the acceptance rate of off-the-shelf LLMs.
