+++ dsi/online/run/README.md
@@ -1,0 +1,1 @@
+This directory implements a DSI Run (see `dsi/types/run.py`) using a thread pool.
