+++ dsi/types/run.py
@@ -1,7 +1,7 @@
 from typing import Generator, final
 
 from dsi.configs.config_run import ConfigRun
-from dsi.offline.common import generate_num_accepted_drafts
+from dsi.offline.run.common import generate_num_accepted_drafts
 from dsi.types.result import Result
 from dsi.utils import set_random_seed
